requirements.txt
pandas
scikit-surprise
codecarbon
psutil
scikit-learn
numpy<2
lightfm